Update README.md
Improved documentation,  use of octicons, emojis. Generally turned this into a project someone might want to fork, and not just by accident:P

This project repository is intended to provide a "skeleton" for you, as a project creator, to provide a more friendly and familiar online experience for the typical end-user, allowing them to simply follow links in your ReadMe.md file to the important stuff. For users previously unfamiliar with GitHub specifically or version tracking in general, simply navigating the page to download the current version of your project or to submit a bug report is difficult.

This project was initially "forked" from http://github.com/holman/feedback, but it has evolved markedly from that point